badge analytics and analytics concern namespace
