+++ app/models/course.rb
@@ -3,7 +3,7 @@ require_relative "role"
 class Course < ActiveRecord::Base
   include Copyable
   include UnlockableCondition
-  include CourseAnalytics
+  include Analytics::CourseAnalytics
 
   after_create :create_admin_memberships
